Fixes validation problem
Calling FullcalendarEngine::Event.new.valid? resulted in
NoMethodError: undefined method `>=' for nil:NilClass on
app/models/fullcalendar_engine/event.rb:21:in `validate_timings'

Starttime and endtime can be nil and '>=' is obviously not available
on nil. Instead, validate_timings now verifies that both start- and
endtimes are set before calling '>='.

Validations for starttime and endtimes are also added, so the
model will still be invalid even though '>=' is not called in
validate_timings.